% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_logbook.R
\name{add_anaesthesia_case_id}
\alias{add_anaesthesia_case_id}
\title{Case Id}
\usage{
add_anaesthesia_case_id(x)
}
\arguments{
\item{x}{\code{data.table}, logbook data}
}
\value{
\code{data.table}, with additional CaseId column.
}
\description{
Add an additional column with a corresponding case id.
}
