% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_logbook.R
\name{case_duration}
\alias{case_duration}
\alias{case_start}
\alias{case_end}
\title{Case times}
\usage{
case_duration(x)

case_start(
  x,
  reference = c("start", "vaporizer-opening", "mechanical-ventilation")
)

case_end(x)
}
\arguments{
\item{x}{\code{data.table}, logbook data}

\item{reference}{\code{character(1)}, reference time point.}
}
\value{
\code{double}, duration in minutes.

\code{POSIXct}, case_start times

\code{POSIXct}, end times
}
\description{
Start/Stop and duration of anesthesia cases.
}
\examples{
f <- system.file("logbook.txt", package = "dltr")
lb <- read_logbook(f)
case_duration(lb)
case_start(lb)
case_start(lb, reference = "mechanical-ventilation")
case_end(lb)
}
