% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_logbook.R
\name{filter_short_cases}
\alias{filter_short_cases}
\title{Filter cases}
\usage{
filter_short_cases(x, min_duration = 10)
}
\arguments{
\item{x}{\code{data.table}, logbook data.}

\item{min_duration}{\code{numeric(1)}, minimal accepted duration in minutes.}
}
\value{
\code{data.table}, filtered cases.
}
\description{
Filter cases
}
\examples{
f <- system.file("logbook.txt", package = "dltr")
lb <- read_logbook(f)
filter_short_cases(lb)
filter_short_cases(lb, min_duration = 60)
}
