% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_logbook.R
\name{is_volatile_anesthesia}
\alias{is_volatile_anesthesia}
\title{Volatile anaesthetics}
\usage{
is_volatile_anesthesia(x)
}
\arguments{
\item{x}{\code{data.table}, logbook data}
}
\value{
\code{logical}, volatile anesthetic used?
}
\description{
Are volatile anesthetics used?
}
\examples{
f <- system.file("logbook.txt", package = "dltr")
lb <- read_logbook(f)
is_volatile_anesthesia(lb)
}
