% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_logbook.R
\name{value_at}
\alias{value_at}
\title{Find values}
\usage{
value_at(
  x,
  label,
  time,
  reference = c("start", "vaporizer-opening", "mechanical-ventilation")
)
}
\arguments{
\item{x}{\code{data.table}, logbook data}

\item{label}{\code{character(1)}, LABEL of value, e.g. PEEP.}

\item{time}{\code{numeric(1)}, time from reference in minutes.}

\item{reference}{\code{character(1)}, reference time point.}
}
\value{
\code{double}, of value \code{label}
}
\description{
Look up values in the logbook for a given time point relative to a reference.
}
\examples{
f <- system.file("logbook.txt", package = "dltr")
lb <- read_logbook(f)
value_at(lb, "FGF", time = 15, reference = "start")
}
